hw4.5: fix2

+++ src/main/java/ru/hogwarts/schoolapi/controller/StudentController.java
@@ -86,7 +86,7 @@ public class StudentController {
     }
 
     @GetMapping("/parallelSum")
-    public int parallelSum() {
+    public long parallelSum() {
         return studentService.parallelSum();
     }
 }
